+++ app/src/androidTest/java/woowacourse/payments/ui/CardNumberTextFieldTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.ui
+package woowacourse.payments.ui
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
@@ -7,42 +7,37 @@ import androidx.compose.ui.test.performTextInput
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.CardOwnerNameTextField
 
-class CardOwnerNameTextFieldTest {
+class CardNumberTextFieldTest {
     @get:Rule
     val composeRule = createComposeRule()
 
     @Before
     fun setUp() {
         composeRule.setContent {
-            CardOwnerNameTextField()
+            CardNumberTextField()
         }
     }
 
     @Test
-    fun `카드_소유자_이름의_길이는_30자를_넘을_수_없다`() {
+    fun `숫자가_아닌_값을_입력할_수_없다`() {
         // when
         composeRule.onNodeWithText("")
-            .performTextInput("모".repeat(30))
-
-        composeRule.onNodeWithText("모".repeat(30))
-            .performTextInput("모찌")
-
+            .performTextInput("123NaN")
 
         // then
-        composeRule.onNodeWithText("모".repeat(30))
+        composeRule.onNodeWithText("123")
             .assertIsDisplayed()
     }
 
     @Test
-    fun `입력된_이름의_길이가_출력된다`() {
+    fun `숫자_4자리마다_구분자가_들어간다`() {
         // when
         composeRule.onNodeWithText("")
-            .performTextInput("모찌")
+            .performTextInput("1234567812345678")
 
         // then
-        composeRule.onNodeWithText("2/30")
+        composeRule.onNodeWithText("1234-5678-1234-5678")
             .assertIsDisplayed()
     }
 }
